requirements.txt
# Deep Learning Framework
torch>=2.0.0
torchvision>=0.15.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
Pillow>=10.0.0
opencv-python>=4.8.0
scikit-image>=0.21.0
scikit-learn>=1.3.0

# Configuration
PyYAML>=6.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Data Augmentation
albumentations>=1.3.0

# Progress Bars
tqdm>=4.65.0

# Data Download
kaggle>=1.5.0

# Experiment Tracking (optional)
wandb>=0.15.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Jupyter Notebooks
jupyter>=1.0.0
ipykernel>=6.25.0

# Code Quality
black>=23.7.0
flake8>=6.1.0

